a27-split-member-service: `src/service/member_service.rs` is 1262 lines
