+++ openspec/changes/archive/2026-05-23-a27-split-member-service/specs/member-service-layout/spec.md
